funcionando Coleção

+++ fabsoft-backend/src/main/java/br/univille/fabsoft_backend/controller/ColecaoController.java
@@ -57,13 +57,17 @@ public class ColecaoController {
         return new ResponseEntity<Colecao>(colecao, HttpStatus.OK);
     }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Colecao> delete(@PathVariable long id) {
-        try {
-            var deletada = service.delete(id);
-            return new ResponseEntity<>(deletada, HttpStatus.OK);
-        } catch (Exception e) {
-            return ResponseEntity.notFound().build();
-        }
+@DeleteMapping("/{id}")
+public ResponseEntity<Void> delete(@PathVariable long id){
+    if(id <= 0){
+        return ResponseEntity.badRequest().build();
     }
+
+    try {
+        service.delete(id); 
+        return ResponseEntity.noContent().build(); 
+    } catch (Exception e) {
+        return ResponseEntity.notFound().build();
+    }
+}
 }
